Rename package to org.apache.pulsar.opentelemetry
