+++ pulsar-opentelemetry/src/main/java/org/apache/pulsar/opentelemetry/OpenTelemetryAttributes.java
@@ -16,4 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.stats;
+package org.apache.pulsar.opentelemetry;
+
+import io.opentelemetry.api.common.AttributeKey;
+
+public interface OpenTelemetryAttributes {
+    AttributeKey<String> PULSAR_CLUSTER = AttributeKey.stringKey("pulsar.cluster");
+}
