+++ pulsar-proxy/src/main/java/org/apache/pulsar/proxy/stats/PulsarProxyOpenTelemetry.java
@@ -22,7 +22,7 @@ import io.opentelemetry.api.metrics.Meter;
 import java.io.Closeable;
 import lombok.Getter;
 import org.apache.pulsar.PulsarVersion;
-import org.apache.pulsar.common.stats.OpenTelemetryService;
+import org.apache.pulsar.opentelemetry.OpenTelemetryService;
 import org.apache.pulsar.proxy.server.ProxyConfiguration;
 
 public class PulsarProxyOpenTelemetry implements Closeable {
